# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel OpenApi Demo Documentation",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "kokodah222@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api/v1",
            "description": "Demo API Server"
        }
    ],
    "paths": {
        "/cart": {
            "get": {
                "tags": [
                    "Carts"
                ],
                "summary": "Get list of user cart",
                "description": "Returns list of user cart",
                "operationId": "getUserCartList",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "403": {
                        "description": "Cart not found"
                    }
                }
            }
        },
        "/cart/add/{product_id}": {
            "post": {
                "tags": [
                    "Carts"
                ],
                "summary": "Store new produt",
                "description": "Create new product",
                "operationId": "storeNewCartProduct",
                "parameters": [
                    {
                        "name": "product_id",
                        "in": "path",
                        "description": "Product id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "403": {
                        "description": "Product not found"
                    }
                }
            }
        },
        "/cart/quantity": {
            "put": {
                "tags": [
                    "Carts"
                ],
                "summary": "cart add product",
                "description": "Cart add product",
                "operationId": "CartAddProduct",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "403": {
                        "description": "Cart not found"
                    }
                }
            }
        },
        "/cart/delete/{product_id}": {
            "delete": {
                "tags": [
                    "Carts"
                ],
                "summary": "cart delete product",
                "description": "Cart delete product",
                "operationId": "CartDeleteProduct",
                "parameters": [
                    {
                        "name": "product_id",
                        "in": "path",
                        "description": "Product id for delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "403": {
                        "description": "Cart not found"
                    }
                }
            }
        },
        "/cart/clear": {
            "post": {
                "tags": [
                    "Carts"
                ],
                "summary": "cart clear",
                "description": "Cart clear",
                "operationId": "CartClear",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Cart error"
                    }
                }
            }
        },
        "/cart/total": {
            "post": {
                "tags": [
                    "Carts"
                ],
                "summary": "cart total",
                "description": "Cart total",
                "operationId": "CartTotal",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "403": {
                        "description": "Cart error"
                    }
                }
            }
        },
...
        "/coupons": {
            "get": {
                "tags": [
                    "Coupons"
                ],
                "summary": "Get list of coupons",
                "description": "Returns list of coupons",
                "operationId": "getCouponsList",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                }
            }
        },
        "/coupons/make": {
            "post": {
                "tags": [
                    "Coupons"
                ],
                "summary": "Store new coupon",
                "description": "Create new coupon",
                "operationId": "couponMade",
                "responses": {
                    "201": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                }
            }
        },
        "/coupons/{coupon}": {
            "get": {
                "tags": [
                    "Coupons"
                ],
                "summary": "Change currency",
                "description": "Change coupon currency",
                "operationId": "changeCouponCurrency",
                "parameters": [
                    {
                        "name": "coupon",
                        "in": "path",
                        "description": "coupon slug",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Coupon not found"
                    }
                }
            }
        },
...
        "/currentValues": {
            "get": {
                "tags": [
                    "Currency"
                ],
                "summary": "change currency",
                "description": "change currency",
                "operationId": "changeCurrency",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                }
            }
        },
...
    },
...
}